# Fix `resize_image_ocv` for `fit_to_window`