# What does this PR do?

<!--
Congratulations! You've made it this far! You're not quite done yet though.

Once merged, your PR is going to appear in the release notes with the title you set, so make sure it's a great title that fully reflects the extent of your awesome contribution.

Then, please replace this with a description of the change and which issue is fixed (if applicable). Please also include relevant motivation and context. List any dependencies (if any) that are required for this change.

Once you're done, someone will review your PR shortly (see the section "Who can review?" below to tag some potential reviewers). They may suggest changes to make the code even better. If no one reviewed your PR after a week has passed, don't hesitate to post a new comment @-mentioning the same persons---sometimes notifications get lost.
-->

<!-- Remove if not applicable -->

Fixes https://github.com/openvinotoolkit/model_api/issues/167

Example source:
```python
import torch
import tempfile
import numpy as np
import matplotlib.pyplot as plt
import openvino
from openvino.model_api.adapters import OpenvinoAdapter, create_core
from openvino.model_api.adapters.utils import resize_image_with_aspect_ocv
from openvino.preprocess import PrePostProcessor

class DummyModel(torch.nn.Module):
    def forward(self, inputs):
        return inputs

model = DummyModel()

with tempfile.TemporaryDirectory() as tmpdirname:
    onnx_path = tmpdirname + "/dummy.onnx"
    ov_path = tmpdirname + "/dummy.xml"
    torch.onnx.export(
        model,
        (torch.zeros(1, 3, 1024, 1024, dtype=torch.float32)),
        open(onnx_path, "wb"),
        input_names=["inputs"],
        output_names=["outputs"])
    ov_model = openvino.convert_model(onnx_path)
    openvino.save_model(ov_model, ov_path, compress_to_fp16=False)
    
    model_adapter = OpenvinoAdapter(
        core=create_core(),
        model=ov_path,
        weights_path=ov_path.replace(".xml", ".bin"),
        device="CPU",
        max_num_requests=1,
        plugin_config={"PERFORMANCE_HINT": "THROUGHPUT"},
    )

    model_adapter.embed_preprocessing(
        layout="NCHW",
        resize_mode="fit_to_window",
        interpolation_mode="LINEAR",
        target_shape=(1024, 1024),
        pad_value=0,
        brg2rgb=False,
    )

img = np.ones((256, 512, 3)).astype(np.uint8) * 255
ov_results = model_adapter.compiled_model(img[None])
np_results = resize_image_with_aspect_ocv(img, (1024, 1024))

plt.figure(figsize=(15,7))
plt.subplot(131)
plt.imshow(img)
plt.title("input image w/ (256, 512)")
plt.subplot(132)
plt.imshow(np_results)
plt.title(f"resize_image_with_aspect_ocv results w/ {np_results.shape[:2]}")
plt.subplot(133)
plt.imshow(list(ov_results.values())[0][0].transpose(1, 2, 0))
plt.title("openvino results w/ (1024, 1024)")
plt.show()
```

Before:
![image](https://github.com/openvinotoolkit/model_api/assets/97221135/49b7d555-40b7-4d05-ac52-7d99f4bbb0f8)

After:
![image](https://github.com/openvinotoolkit/model_api/assets/97221135/1c48719e-b42f-4607-bd5b-7f18b85dac81)


## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?