@sovrasov 2024-04-10
@sungchul2 thanks the PR looks good. Could you add the reproducer from the PR description to `tests/python/precommit/` as a unit test? I believe we could check that `sum(abs(ov_image - np_image)) < error`